refactor(chat): isolate index fingerprint validation
